feat(*): Adds basic caching support and fetching for the CLI client
Note that I only implemented the cache part for the new subcommand.
I'll add caching to the other commands in a follow up PR. There are
a bunch of code changes here. Most of them were "campfire rule" cleanup
as I found some bugs and inconsistencies, as well as some restructuring
to make things cleaner