ref(client): Updates the create_parcel_from_file method
This has 2 specific changes. The first is that it changes the signature
to take a label object rather than a path to a label. After consuming the
client module in the CLI, I discovered that it is much more common to be
building the label in memory from various configuration flags or available
data, so having to write it to a file path would have made no sense.

The second is that I essentially disabled the function. Because this does
not work without a fix we merged into reqwest, it now just returns an error.
This is also reflected in the documentation. Once reqwest cuts a new release
we can do the same