ref(storage): Updates the storage trait to use streams
As we have gone along with client implementations, I realized that we
were always converting things from streams into readers. Because this
is an HTTP API, it should be expecting an arbitrary stream of data from
somewhere rather than a reader of some kind.

This also fixes some clippy lints I found