fix(provider/file): Uses partfile for invoice and improves locking
This has some small improvements to the new `PartFile` to make sure that it
takes a file "lock" right when it is created. It also refactored the `create_invoice`
method to use the new `PartFile`. I also added a test to at least give
us a sanity check that a double write doesn't cause an issue