ref(server): Adds test to make sure signed invoice is returned
I noticed when removing the clones that we weren't signing the invoice passed in
to the `create_invoice` function. This meant that the server would be returning
an unsigned invoice even though it was actually signed properly. This commit adds
a test that checks for proper host signing by the server