feat(proxy): Adds support for signing and verifying as a proxy
As part of this, I realized that `Proxy` isn't the best thing to be using
for the client because it isn't truly a Proxy as it is acting as a provider
wrapper. This meant that we'd have a bunch of verification and signing code
you could turn off when using the proxy. Instead of doing that I just
did a `Provider` implementation for `Client`. This is a little less DRY, but
better separation of concerns and a cleaner API for `Proxy`

Closes #141