feat(*): Adds new `Verified` and `Signed` traits
This changes how the providers work by only taking invoices that have
been verified and signed. This is accomplished using a sealed trait
so that signing must be done by library code. The big change is that
providers no longer have to implement signing.

This work supersedes (and contains the commits from, for history sake
PR #168

Closes #141