feat(*): Adds `cargo deny` to build pipeline
This adds package auditing and `cargo deny` to the build workflow for
bindle. This should inform us if we are using any projects with security
advisories or ones we have explicitly denied (none right now). This will
also run a license check to make sure we are at least using licensed
projects. It can be tweaked to allow specific licenses in the future.

As part of this, I have addressed any changes it found