fix(tests): Attempts to fix macos flake
I believe that the slowness of the GH runners was creating issues when using
`cargo run` to run the server binary - even after previously building it. This
changes things to run the built binary directly instead, avoid the overhead of
"rebuilding" it. It also makes things more generic so people consuming test
utils are not required to have a specific binary name