exp(provider): Adds support for experimental embedded provider
This PR adds support for a new embedded `Provider` backed by the `sled`
embedded database and with more efficient data encoding. This new
provider is behind a feature flag and is not compiled into `bindle-server`
by default. Tests have also been updated to check both providers, but
the client tests still only test with the `FileProvider`.

The motivation behind this is that the file system provider requires a bunch of
bespoke logic to check things like locking and various OS specific things. We
also have to front everything with an in-memory LRU cache. Tools like `sled`
already have most of these features built in, plus additional guarantees from
being a database (such as atomic transations and upsert commands). This also
uses a binary encoding format to store the data using less bytes and to
decrease serialization and deserialization time. All of these reasons
sounded like pretty good justification for giving it a whirl. Use of bindle
can continue as before until we decide whether we like this or not as it
is all behind a feature flag.

Any comments on open TODOs are welcome!