feat(*): Adds OIDC login functionality to the server and the client
Authentication is now done using OIDC and JWT verification for completely
stateless auth. This is a major feature add with several breaking changes.
In particular, to construct a client you now need a `TokenManager` to
properly provide a token (if any) to your requests. This ended being much
cleaner than handling several different options

Note, there are probably various refactors that could be done here and those
are noted with TODOs. Most of them are things that could change anyway as
we add further authn/z features.

To test this feature, all you need is the client ID for any OIDC compatible
provider. For example, there is a common one for using Microsoft accounts
for anyone with a public Microsoft account. You can also create your own
Oauth application in something like AAD.