feat(cli): Adds support for standalone tarballs to client
Namely, the `get` command when using export will now output a tarball rather
than a directory. The `push` command now can use a directory or a tarball and
will infer that automatically from the file extension. I also added a new
`package` command that will package up a standalone bindle directory as a
tarball. An `expand` command was considered but rejected for the reason that
there is no special tooling required outside of `tar`. If we wish to add this
in the future, that is totally fine