feat(*): Enforce signatures on server
This completes the change to full enforcement of signing and verification on
both client and server. This PR specifically is for the server. It will no
longer accept bindles that do not have a signature. All signature rules
are enforced according to the verification strategy. Please note that most
of the changes in this PR were to update the tests, which was a good exercise.

This is a breaking behavior change that should be well-documented at the next release.

Closes #106