ref(*): Better organizes features and dependencies
Sometimes a dependent project may need just the types, so now when
no features are enabled, just the core types and traits are exposed.

This also takes a first pass at "optionalizing" as many dependencies
as possible. The dependency tree has gotten quite large, so this tries
to only import things when they are absolutely needed. We could probably
add a few more features (like on the auth stuff) to make it perfect, but
this was a good start