fix(cargo): Updates dependency tree to avoid a compile error on tokio_util
For some reason, if you run a `cargo update` right now, the use of
`tokio_util::codec` stops working. Looking at the code, I had no idea how
it worked in the first place, but now it should work with any combination
of the correct versions. This will be a 0.8.1 patch release to fix the
library

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>