fix(wasm): Fixes deps for building in wasm
A while ago we made sure that if you wanted to pull in the base bindle types
into something compiling to wasm, it would build. Recently, tokio made a
change that made it so it wouldn't compile for wasm with the `fs` feature
enabled. This makes sense, but I did have to untangle some fs stuff from
the signatures module.

While I was here I fixed a bunch of new clippy lints from the latest version
of clippy

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>